+++ IronSource/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 7.2.1.0.1
+* Add support for passing in a presenting view controller.
+
+## 7.2.1.0.0
+* Certified with IronSource SDK 7.2.1.0.
+
 ## 7.2.0.0.0
 * Certified with IronSource SDK 7.2.0.0.