+++ MyTarget/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 5.15.0.2
+* Add support for passing in a presenting view controller.
+
+## 5.15.0.1
+* Add support for Binary CCPA.  
+
 ## 5.15.0.0
 * Certified with MyTarget SDK 5.15.0.