+++ Verve/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 2.11.1.2
+* Add support for passing local parameter "is_location_collection_enabled" to set `+[HyBid setLocationUpdates:]`.
+
+## 2.11.1.1
+* Add support for location updates.
+
+## 2.11.1.0
+* Certified with Verve SDK 2.11.1.
+
+## 2.11.0.1
+* Update user consent to not override existing GDPR and CCPA privacy strings.
+
+## 2.11.0.0
+* Certified with Verve SDK 2.11.0.
+
 ## 2.10.0.0
 * Certified with Verve SDK 2.10.0.
 * Update open source versions to allow compilation with AppLovin SDK v11.0.0+.